Restructured functions to fill out forms into a FormFiller js class.
More importantly, input fields are now matched via regex expressions. This means we dont look for <input id='super-specific-email-id'> on a page, but test for inputs that have an id, placeholder, aria-label, or class that matches a regular expression indicating an email field. Currently this only supports text fields, but radio buttons and dropdowns will be implemented next.