Added a very specific database table used for caching search queries (not a Django database cache, but a specific table).
This maps odd queries to city objects.

My issue involved specific parsing. For example, I would often get cities with names like "CASantaclaravalley, <No State>" and derivatives of that. It was a nightmare to parse into the <Santa Clara, California 85050> city object. The idea is that I will try to parse a querey, and if I can't I will have the Google maps api do it for me. If we succeed in finding a city, state, and zip combination out of an odd querey we will cache it.